The plugin now also shows when creating new wiki pages. Fixes #1
The plugin was previously only inserted on the edit action of the
WikiController. It is now also inserted on the show action, which
ChiliProject uses to create new wiki pages and to show them. This does
mean that for now that plugin also gets loaded when the wiki page editor
is in fact not available