Resizing the editor area resizes the preview
On some browsers it is possible to resize the editor textarea. Such
changes should be reflected in the preview window.

This behavior is now implemented by listening to and acting on the
mouseup event on the textarea. The current implementation is a
workaround workaround since the resize event does not work on individual
HTML elements